# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Documentation",
        "contact": {
            "email": "ivan@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Api Documentation"
        }
    ],
    "paths": {
        "/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get All users with count per page",
                "operationId": "usersAll",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "The number item per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status ok"
                    },
                    "500": {
                        "description": "server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create new user",
                "operationId": "createUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "200": {
                        "description": "answer",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "204": {
                        "description": "sucess"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "UserStoreRequest": {
                "title": "User storring request",
                "description": "User model",
                "properties": {
                    "name": {
                        "title": "Name",
                        "description": "Some text field",
                        "type": "string",
                        "format": "string",
                        "example": "test"
                    },
                    "email": {
                        "title": "Email",
                        "description": "Some email",
                        "type": "string",
                        "format": "string",
                        "example": "test@gmail.com"
                    },
                    "photo": {
                        "title": "Photo",
                        "description": "Some photo",
                        "format": "mimes"
                    },
                    "password": {
                        "title": "Password",
                        "description": "Some pass",
                        "type": "string",
                        "format": "string",
                        "example": "123445EE"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "token": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "User",
            "description": "You can add and see users here"
        }
    ]
}